pyproject.toml
[tool.poetry]
name = "audit-trail"
version = "2.2.4"
description = ""
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.10"
elasticsearch = "7.15.2"
fastapi = "^0.112"
fastapi-utils = "^0.7"
gunicorn = "20.0.4"
httptools = "0.6.1"
httpx = "^0.23"
opentelemetry-api = "1.11.1"
opentelemetry-exporter-jaeger = "1.11.1"
opentelemetry-instrumentation = "0.30b1"
opentelemetry-instrumentation-elasticsearch = "0.30b1"
opentelemetry-instrumentation-fastapi = "0.30b1"
opentelemetry-instrumentation-httpx = "0.30b1"
pydantic = "^1.9"
PyJWT = "^2.6"
python-dotenv = "0.19.1"
python-multipart = "0.0.9"
uvicorn = "0.12.3"
uvloop = "0.20.0"
fastapi-health = "^0.4.0"
pilot-platform-common = "^0.2"

[tool.poetry.dev-dependencies]
async-asgi-testclient = "1.4.9"
pytest = "7.1.1"
pytest-env = "0.6.2"
pytest-asyncio = "0.18.1"
pytest-httpx = "0.21.0"
pytest-mock = "3.7.0"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
